Add NPOI package for parsing Excel files
ExcelDataReader package was removed, because it only reads and doesn't write .xlsx files. Fortunately, NPOI can do both things.

There is a small snippet of code which reads the Films2.xlsx, fills dtTable and after that adds the data to filmsDataGridView.

Besides that the BalloonTip of tray icon was added.